pyproject.toml
[project]
name = "customer-survey"
version = "0.1.0"
description = "Customer Survey Application"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.0.0",
    "flask-login>=0.6.3",
    "flask-sqlalchemy>=3.1.1",
    "flask-wtf>=1.2.1",
    "pydantic>=2.12.5",
    "python-dotenv>=1.0.0",
    "werkzeug>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "pylint>=3.0.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
]
